feat(data-service): split DataService into interface/implementation (#3010)
Split the `DataService` class into a `DataService` interface and
a `DataServiceImpl` implementation class, for better separation
between our public API that the components can rely on (and that
all other data source providers, e.g. Data Explorer, need to
implement as well) and the specific implementation that we use
in Compass.