fix(data-service): use dependency injection for getting ClientEncryption
The mongodb-client-encryption package gets its corresponding driver
package through `require('mongodb')` by default. However, when packaging
Compass, we include mongodb-client-encryption in the unpacked asar,
but not the driver itself, so this fails.

For these cases (and also just to generally ensure that versions match),
an explicit way to specify the driver package is available,
so we should use that here.