fix(connection-form): delete KMS provider option when empty COMPASS-5798
Instead of setting the KMS provider’s config to `undefined`
when no options are provided for that KMS, just delete it entirely.

In particular, previously connectivity would have been
impacted by the fact that `hasAnyCsfleOption()` would
not be able to properly handled `undefined` values in the
KMS provider config. That function is also adjusted here
to handle these cases more defensively.