chore(data-service): track equality-searchable encrypted fields
In addition to tracking all encrypted fields for a given collection’s
schema sources, also track equality-searchable encrypted fields.
This is helpful for knowing whether a field can be included in
the query part of a Compass CRUD update operation that includes
original values of fields in the query portion of that update.

As part of this, the list of encrypted fields reported by
`knownSchemaForCollection()` and used internally in the encrypted
field tracker is changed from a simple array of strings to a
pair of field path lists, one for all encrypted and one for
equality-searchable encrypted fields in the schema. Using
field paths instead of `.`-joined strings has no direct effect
here, but aligns this part of Compass with the recent changes
to nested field editing/dots and dollars support.