fix(compass-crud): exclude unindexed encrypted fields in update query COMPASS-5809
When updating unindexed encrypted fields, exclude them from the
find part of the update operation, since they are by definition
not queryable.

This modifies the HadronDocument query building methods to take
options arguments instead of key lists, and updates these options
to be passed as field path lists rather than a plain array of strings
(where `.` was ambiguous between literal field names and nested
fields, which was not a problem previously because the field paths
were always coming from the context of shard keys, where `.` is
unambiguous but now is with FLE1 nested fields).