chore: add BSON doc editor component, use for FLE2-related fields COMPASS-5949
This partially addresses COMPASS-5949 and might be useful as a starting
point for what approach we want to take here.

Currently, this creates a new `BSONDocumentEditor` component
specifically for editor instances that are targeting BSON documents
(which most of our editors do), adds an EJSON/Shell syntax switch
to it, and centralizes parsing/JS serialization in a single location.