fix(compass): include node-forge in unpacked node_modules COMPASS-5802
Since it is a dependency of win-export-certificate-and-key,
node-forge should also be unpacked, like its dependent.

This adds an e2e test for verifying that the shell works
when connecting to Atlas using tlsUseSystemCA (since that
is an environment which always uses TLS and should always
match the system CA suite).