chore(compass): re-reverse identify()/track() order in CompassTelemetry
6fa86474c8e4 introduced a call to `getOsInfo()` in
`CompassTelemetry.identify()`. This accidentally changed the order
of calls to be `track()` called before `identify()` on the segment API,
since `.identify()` would be called asynchronously.

I’m not sure if this has a noticable impact on Segment, but we
generally should be calling `.identify()` first I believe.

Also, use this as an opportunity to log errors instead of
entirely ignoring them.