feat(compass): add CLI connection import/export COMPASS-6073 COMPASS-6074 (#3454)
* chore(compass-e2e-tests): allow spawning one-shot compass processes

* chore(connection-model): allow accessing from main process

`splice-disk` is the main process equivalent of `splice-disk-ipc`
for renderers. Properly expose it and use it in connection-model.

* feat(compass): add CLI connection import/export COMPASS-6073 COMPASS-6074

Add support for exporting and importing favorite connections when
using Compass as a CLI. Users can specify `--export-connections=<file>`,
`--import-connections=<file>` and optionally also a passphrase
through `--passphrase=<string>`.

The encryption/decryption and export/import logic are added to the
data-service package. The Compass main entry point is modified
to the degree to which it is necessary to allow running it as
a CLI rather than always in GUI mode.