fix(compass-preferences-model): tell yargs about types for kebab case options
Yargs is informed about the types of options where they can be determined, and
performs camel-case expansion on options specified in kebab case.

However, it does not apply those types in the latter case, making it so that
`--showed-network-opt-in foo` would lead to a `false` value for `showedNetworkOptIn`,
unlike `--showedNetworkOptIn foo`. Address this by also telling yargs about the
kebab-case versions explicitly.