fix(compass-settings): keep internal state of settings modal up-to-date COMPASS-6296
Refactor the settings modal internals so that it keeps a “sandbox”
storage of preferences that are not store directly, but have the same
behavior as the actual preferences storage, instead of manually
tracking settings changes.

This way, we can ensure that we can display the state of a setting
without actually having to save it (in the specific case that requires
this is the `enableShell` / `readOnly` pair, where clicking `readOnly`
should immediately indicate that setting it also changes `enableShell`.