chore: generate ts definitions for more packages (#3798)
Add `tsc --emitDeclarations` to all packages that use `webpack-compass`
in a way that is identical to what we currently do in
`compass-aggregations`. This helps with getting proper type definitions
when using these packages in other packages (and this PR includes two
small type fixes that were made necessary by existing mismatches).