feat: add --show-example-config flag COMPASS-6084
The original config file suggested including a file in the packages
themselves. However, this is hard for a number of reasons; for example,
our .deb and .rpm generators do not provide options for specifying
configuration files, and if they did, config files would conflict
between the different variants of Compass. For the .zip distribution,
we ship the Compass executable at the top level of the .zip, meaning
that it is impossible to actually write to the path where Compass
expects the configuration file just by unpacking the .zip.

As a lightweight alternative, Compass now supports a new
`--show-example-config` command line flag, and refers to it, together
with a full command showing its usage, in the `--help` output.