fix(compass-preferences-model): avoid race condition when installing listeners
The following sequence of events is currently possible:
- A preference value is changed in the main process
- The main process informs the renderer about it
- Consequentially, the renderer starts to refresh its cache
- Something (X) running in the renderer gets the preference value
  via `getPreferences()` and then listens for changes using
  `onPreferenceValueChanged()` (as e.g. the React preference utils do)
- The cache refresh operation finishes and updates the cache

X then uses an outdated value of the preference and is not informed
about the change in value.

Fix this by keeping track of active cache refresh operations, and
if there are any, call the newly installed listener once with
the fresh value.