fix(compass-shell): let enableShell setting control runtime start/stop
Make it so that the `enableShell` toggle does not only control whether
the shell is displayed, but actually create or tear down the shell
runtime instance in according to it.

This obviously does not fix any issues on its own, but it does unblock
development on e.g. https://github.com/mongodb-js/compass/pull/4484,
and seems like a good idea in general, because it also gives users
a workaround around not just the current Electron bug, but also future
problems where having the shell process running in the background
leads to undesirable behavior in the main Compass app.