chore: avoid process.nextTick() and setImmediate() COMPASS-7314 (#4974)
Replace, or where unused remove, usage of `process.nextTick()`
and `setImmediate()`, both of which are Node.js-specific APIs
for async deferral for which browser-compatible replacements with
near-identical behavior exist.

Add linting rules to prevent usage of these methods. There might be
a way to scope these linting rules to only the packages that we
intend to use in browser environments, or exclude them from applying
to test files. However, that seems unnecessary, given that the
replacements are no harder to use, so a consistent style across
the entire Compass codebase seems preferable.