feat(release): write artifacts to new release bucket with private acls COMPASS-9238
This commit adjusts our release process to upload artifacts to the new
release bucket using private ACLs. Because users fetch artifacts from
our CDN, we are not allowing direct access to this bucket. If we do not
set the ACLs to private, our uploads will fail. In order to support this,
we had to update to the latest devtools-shared dl-center package.