chore: remove Atlas Sign-In logic COMPASS-9595
Since we're removing the signin requirement from the AI features we do not hacve any components in need of the Atlas sign-in flows. As it's also built using a pattern we have deviated from, it's best to remove it now and re-add it through a different structure later if needed.